FossilOrigin-Name: e65e6c01d632beec9a21758cce6f7ac2e3bc2c5fac88b3bb393b9040fb2cbea6

+++ ext/userauth/userauth.c
@@ -40,7 +40,7 @@ static sqlite3_stmt *sqlite3UserAuthPrepare(
   char *zSql;
   int rc;
   va_list ap;
-  int savedFlags = db->flags;
+  u64 savedFlags = db->flags;
 
   va_start(ap, zFormat);
   zSql = sqlite3_vmprintf(zFormat, ap);
